requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
pymysql==1.1.0
cryptography==41.0.7
mysql-connector-python==8.1.0
flask-mail==0.9.1
python-dotenv==1.0.0
reportlab==4.0.7
openpyxl==3.1.2
google-generativeai==0.3.2
